macOS: Extract framework creation to sky_utils.py
This refactors create_mac_framework.py to extract framework creation to
sky_utils.py. The only other changes are minor variable renaming and
extraction of functions to make the code more readable/easier to work
with.

The resulting zip archives have been verified to be identical before and
after.

This is refactoring to support embedding dSYM debug information in a
follow-up patch.