clangd_check: write .clangd as part of test
This eliminates the checked-in `.clangd` file at the root of the engine
repository, which gets picked up by editors that rely on `clangd`. In
particular, the `Remove: [-m*, -f*]` is problematic for iOS/macOS
embedder developers since it removes the `-fobjc-arc` flag, which causes
`clangd` to emit errors when using ARC features. Example:

   Cannot create __weak reference in file using manual reference counting

This removes the checked-in file, and instead writes the file before
running the test, then cleans it up afterwards. Unfortunately, `clangd`
does not appear to have a mechanism to point to a config file elsewhere
and run as though it were present in a specified directory.

Fixes: https://github.com/flutter/flutter/issues/154064