macOS: Do not archive/upload FlutterMacOS.dSYM to cloud
As of the following patches, we now bundle FlutterMacOS.framework.dSYM as part of FlutterMacOS.xcframework. The dSYM is automatically copied into the release build products directory, and bundled in the .xcarchive produced by Xcode's *Product > Archive* feature which produces bundles for upload to the App Store.

* https://github.com/flutter/engine/pull/54696
* https://github.com/flutter/flutter/pull/153975

The .dSYM bundle is now available both in the uploaded .xcarchive and in the xcframework in Flutter's internal artifact cache. For developers with CI toolchains that do additional manual handling or local archiving of .dSYMs, the dSYMs no longer need to be downloaded from cloud storage as previously detailed in `docs/Crashes.md`, but can instead be copied up from the appropriate dSYM subdirectory in the framework cache:

* `flutter/bin/cache/artifacts/engine/darwin-x64-release/FlutterMacOS.xcframework`

Also adds documentation for crash symbolication on macOS.

Issue: https://github.com/flutter/flutter/issues/153879