iOS,macOS: Add logging of duplicate codesign binaries
When duplicates are found in the input lists, log them to stderr. For
the case of duplicates across list, we print the whole list, sorted,
rather than three separate lists, as it makes it simpler to spot the
duplicate, at which point the file will be easy to spot in either
`create_ios_framework.py` or `create_macos_framework.py`.