[multiple] Avoid new `unreachable_switch_default` warning.
The Dart analyzer will soon be changed so that if the `default` clause
of a `switch` statement is determined to be unreachable by the
exhaustiveness checker, a new warning of type
`unreachable_switch_default` will be issued. This parallels the
behavior of the existing `unreachable_switch_case` warning, which is
issued whenever a `case` clause of a `switch` statement is determined
to be unreachable.

Before adding the new warning to the analyzer, code in the engine
needs to first be updated to eliminate these unreachable `default`
clauses, so that the warning won't cause builds to break.

For more information, see
https://github.com/dart-lang/sdk/issues/54575.