Match Windows logic for picking RGB/BGR textures.
The existing code had a lot of additional checks that didn't seem to need to be there.

BGR could be passed back to Flutter, but this was never used in creating the texture.

There has been a report of a Flutter app on Linux with swapped red and blue
color channels, so this seems like it is likely not working on some drivers.

The original logic was introduced in f104bad4ca68bf4cd87a2ab83cb32d82fe0dea8b