[Impeller] Ignore opt-outs on iOS devices.
Addresses https://github.com/flutter/flutter/issues/155541

This does **not** remove Skia itself from the build. I'll stage that in
a followup. Want to keep the scope of this patch small.

This is the new behavior:

* On iOS hardware **and** simulators, the engine flag (`--enable-impeller`) to toggle Impeller will be ignored.
* On iOS hardware **and** simulators, the PList flag (`FLTEnableImpeller`) to toggle Impeller will be ignored.
* On iOS hardware **and** simulators, the default renderer is Impeller.
* On iOS simulators **only**, running on emulated environments without access to a Metal device will fallback to Skia with software rendering.

The settings objects enable_impeller field is not static constexpr to avoid accidentally disabling the flag while it still exists.