On a merged UI/platform thread, immediately execute platform message handlers and then run microtasks
Shell::OnPlatformViewDispatchPlatformMessage is called on the platform thread
and must invoke the message handler on the UI thread.  If the platform and UI
threads are merged, then the handler can be run immediately.  The shell can
then post an empty task to the UI thread to make the UI message loop run the
task observer that drains the microtask queue.

Fixes https://github.com/flutter/flutter/issues/156595