iOS: Eliminate needless profiler metrics ivar
iOS profiling is implemented in ProfileMetricsIOS, which has a default
constructor, non member fields, and no superclass. i.e. it's roughly
zero cost to create and holds no state. There's therefore no reason to
make it an ivar on FlutterEngine.

Further, making it an ivar on FlutterEngine means that the profiling
sampler lambda needs to capture self which, in the case where all other
FlutterEngine references go out of scope during sampling, means that the
last FlutterEngine reference exists on a profiling thread, and when the
sampling lambda block completes, the engine is dealloc'ed.

The engine *must* be deallocated on the platform (main in Apple
terminology) thread because FlutterEngine holds a reference to a
PlatformViewsController, which owns a WeakPtrFactory. WeakPtrFactory's
dtor asserts that it executes on the thread on which it was created, in
this case, the platform thread.

Since there's no need for ProfileMetricsIOS to be tied to the engine, we
now create it on the fly in the sampler.

No test changes because no semantic changes.

Uncovered by ARC migration.

Issue: https://github.com/flutter/flutter/issues/137801
Issue: https://github.com/flutter/flutter/issues/156177