iOS: Improve thread safety of first frame callback
Previously, we dispatched a task to a background thread to wait for the
first frame to be ready. In the background task, we then throw a task on
the main thread that invokes that user-provided callback to notify them
that the first frame is ready.

During ARC migration, we ensured that self was strongly-captured in the
task that runs on the main thread, in order to prevent the possibility
that the last remaining reference to the FlutterEngine instance be the
one on the background thread.

However, with the previous code, it's possible that the callback is
dispatched to the main thread, executes, and completes before the block
on the background thread finishes. In that case, the last remaining
FlutterEngine reference may *still* be the one on the background thread.

In order to prevent this scenario, we use `dispatch_groupd_notify` to
ensure that the tasks are executed sequentially, and that the callback
task is not dispatched to the main thread until the background task has
completed. Further, we capture FlutterEngine strongly in the second task
to ensure that *it* is the last remaining task.

No changes to tests since this isn't testable as written.

Issue: https://github.com/flutter/flutter/issues/156177