iOS: Refactor ShellTestPlatformViewMetal
Migrates DarwinContextMetal to a plain C struct, eliminating the need
for constructor, getters, etc. since it's only used in this translation
unit. The fields themselves cannot be inlined as fields on
ShellTestPlatformViewMetal because the header in which that class is
defined is included in plain C++ (non-Obj-C++) translation units and
therefore cannot contain Obj-C types.

This simplifies the code and simultaneously fixes complicated ARC
behaviour in which the const "DarwinContextMetal::context()" getter
caused refcount to be incremented on the underlying context_ pointer,
but not decremented even if never unassigned. In particular the line
```objc
FML_CHECK([metal_context->context() mainContext]));
```
appeared to cause refcount to be incremented but never decremented.

Regardless of the ARC issue, this significantly simplifies the code.

This also eliminates the last remaining use of fml::scoped_nsobject in
Flutter's codebase. That class will be removed in a followup PR.

Issue: https://github.com/flutter/flutter/issues/137801