[Impeller] Delete any remaining GL objects during destruction of the ReactorGLES
At shutdown time the ReactorGLES may still be holding handles of GL
objects.  These objects should be cleaned up when the reactor is deleted.

This leak can be seen by running DlGoldenTest.ShimmerTest, which takes
a series of screenshots.  Each screenshot creates an AiksContext.
Without this change, the textures in the AiksContext's ReactorGLES will
be leaked after the AiksContext is destroyed.