iOS,macOS: Add Obj-C cflags to all Obj-C targets
Ensure that all Objective-C code in the codebase are being built with
the standard set of Flutter Objective-C compiler flags with ARC enabled.

Also bumps the cflags config up to the top of the first block within
each target in which Objective-C sources appear, so that the location is
consistent.

Issue: https://github.com/flutter/flutter/issues/137801