iOS: Use fml::CFRef in place of Scoped
`fml::CFRef` implements the bulk of the operations implemented in the
one-off `Scoped` class. It doesn't implement the `handle()` method that
allows direct writing into the internal storage of the wrapper, but that
method is effectively an escape hatch for all the safety guarantees
provided by the wrapper, so it seems safer to avoid adding it.

Issue: https://github.com/flutter/flutter/issues/137801