macOS: ARC bridge casts for FlutterMetalTexture.user_data
In `[FlutterSurface asFlutterMetalTexture]` we return a
`FlutterMetalTexture` whose `user_data` field holds a void* reference to
the associated `FlutterSurface`.

For conistency with other parts of the codebase, this now uses ARC
bridge casts to perform the additional retain and release of the
associated surface rather than CoreFoundataion functions
`CFBridgingRetain` and `CFBridigingRelease`.

Also migrates the code to initialise and immediately return the struct
using field designators, and reorders the assignments to be in the
declaration order specified in embedder.h as required by the C++ spec.

No changes to tests since this introduces no semantic change.

Issue: https://github.com/flutter/flutter/issues/137801