[CP-stable]Add a check for the surface if it is valid (#56301)
This pull request is created by [automatic cherry pick workflow](https://github.com/flutter/flutter/blob/main/docs/releases/Flutter-Cherrypick-Process.md#automatically-creates-a-cherry-pick-request)

### Issue Link:

https://github.com/flutter/flutter/issues/155018

### Changelog Description:

Summary:

A crash in Flutter SDK 3.24.1 affects Android apps using PlatformView, not limited to Google Maps. The crash occurs when the surface is released unexpectedly, causing subsequent draw calls to fail.

### Impact Description:

**Summary:**  
A crash in Flutter SDK 3.24.3 affects Android apps using `PlatformView`, not limited to Google Maps. The crash occurs when the surface is released unexpectedly, causing subsequent draw calls to fail.

**Impact:**  
This issue can lead to app crashes on Android for any `PlatformView` usage, affecting stability across a range of widgets and plugins that rely on native views.

### Workaround:
Is there a workaround for this issue?

Release the `Surface` in a regular way might avoid the issue, per https://github.com/flutter/flutter/issues/155018#issuecomment-2363602469

### Risk:
What is the risk level of this cherry-pick?

### Test Coverage:
Are you confident that your fix is well-tested by automated tests?

### Validation Steps:
What are the steps to validate that this fix works?

- https://github.com/alex-kobozev-talabat/flutter_surface_example
- https://github.com/flutter/flutter/issues/155018#issuecomment-2363602469