iOS,macOS: Build/test test_metal_surface_unittests
Years ago, chinmaygarde@ created TestMetalSurface and wrote a unit test for it in ed30d77ab9d16706829adc8214d1fe1949696d5d.

And some things that should not have been forgotten were lost. History became legend. Legend became myth. And for four and a half years, the test passed out of all knowledge. Until, when chance came, it ensnared a new bearer. The test came to the creature cbracken@, who took it deep into the tunnels of the GN build files. And there it consumed him. The test brought to cbracken@ unnatural long life. For 500 hours, it poisoned his mind. And in the gloom of cbracken@'s cave, it waited. Darkness crept back into the forests of the world. Rumour grew of a shadow in the East - whispers of a nameless fear. And the test of unit perceived its time had now come.