[Impeller] libImpeller: Reset the GL state when transitioning control back to the embedder.
Impeller is resilient to OpenGL state being trampled upon when accessing the GL
context. But the embedder may not necessarily be. Ideally, we'd be using saving
the state and restoring it. But that might be too involved. For now, this sets
the GL state is a sane "clean" state.

We could, in theory, do this after each render pass but that unnecessarily
increases API traffic. For now, I have added it at the transition of the
embedder boundary.