Move SetRenderTargetType to EmbedderTestCompositor
SetRenderTargetType is used to configure the backingstore producer on
the compositor, but the backingstore types available to any given
compositor are limited to the specific graphics backend in use:
Software, GL, Metal, or Vulkan.

This moves SetRenderTargetType to EmbedderTestCompositor and its
subclasses and adds RenderTargetType validation. A follow-up patch will
refactor EmbedderTestBackingStoreProducer into backend-specific
subclasses.

For OpenGL backingstore producers, the egl_context_ from
EmbedderTestContext (which is actually set in the EmbedderTestContextGL
subclass and should live there, but that's a separate cleanup) is
required in SetRenderTargetType, so we now inject it into
EmbedderTestCompositor in the constructor so it's available when needed.