[Impeller] Allow per call debug error checking and logging.
In the OpenGL backend, the following behavior now applies:

If IMPELLER_DEBUG is unset (release runtime mode only):
* Zero overhead during GL function dispatch.

If IMPELLER_DEBUG is set (debug and profile runtime modes only):
* Debugging options are checked for error checking, call logging, and thread checking. Except thread checking for UseProgram, all options are false out of the box.
* Options for individual procs or the entire table can be toggled at runtime.

If IMPELLER_DEBUG is set and UNOPT modes are enabled:
* Debug error checking is enabled out of the box. This is Aarons preference and matches previous behavior.

The difference between this and what is on ToT is that debugging is now available in OPT modes but the flags that enable checking are off by default (in OPT modes).

Wires up debugging options for the libImpeller.