iOS: Eliminate FlutterEngine dealloc notification
FlutterEngineGroup keeps an array of all live FlutterEngine instances it
has created such that after the first engine, it can spawn further
engines using the first.

Previously we manually managed this array by adding engines to it upon
creation, then having [FlutterEngine dealloc] emit a notification such
that FlutterEngineGroup can listen for it, and remove instances from the
array upon dealloc.

Instead, we now use an NSPointerArray of of weak pointers such that
pointers are automatically nilled out by ARC after the last strong
reference is collected. This eliminates the need for the manual
notification during dealloc.

Unfortunately, NSPointerArray is "clever" and assumes that if the array
has not been mutated to store a nil pointer since its last compact call,
it must must contain a nil pointer and can thus skip compaction. When
holding weak pointers under ARC, this is no longer the case and so we
work around the issue by storing a nil pointer before calling compact.

See: http://www.openradar.me/15396578

I'm not thrilled with the fact that we're replacing one sort of TODO
with another, but the code is much simpler and avoids relying on a trip
through the notification center, which seems objectively better.

Issue: https://github.com/flutter/flutter/issues/155943