iOS: Eliminate unnecessry use of NSClassFromString
When performing `isKindOfClass` checks, we were occasionally looking up
the class in question using `NSClassFromString()`, instead we now check
against the class directly, which has the added benefit of being
type-safe, and not succeptible to string typos.

Moves the declaration of ForwardingGestureRecognizer and
FlutterDelayingGestureRecognizer to the FlutterPlatfomViews_Internal.h
header, which is non-public.