iOS,macOS: Fix string constant declarations
A couple string constants had been incorrectly declared `const
NSString*` rather than `NSString* const`. The former is a pointer to an
immutable string, but (a) the pointer can be reassigned and (b) the
default NSString implementation is immutable. The latter is an immutable
pointer to a string, which is what we want, since we don't want
developers reassinging our constants to other values.

Also updates an identifier name to pass lint rules.