iOS: Eliminate unnecessary @synthesize directives
In modern Objective-C, `@property` directives automatically generate a
backing ivar (property name prefixed with an underscore), the getter,
and (for readwrite properties) the setter. `@synthesize` directives are
generally only required if the backing ivar has a different name from
the property.

Also updates the FlutterMetalLayer API to match CAMetalLayer:
* Adds API_AVAILABLE declaration to match that on CAMetalLayer.
* Eliminates wantsExtendedDynamicRangeContent property as it's also part
  of CALayer's interface and unused in our implementation.

Previously, we were overriding the behaviour of
UIViewController.prefersStatusBarHidden by synthesizing
_flutterPrefersStatusBarHidden as a backing ivar. Since we're explicitly
overriding the behaviour of a superclass, it's more idiomatic to
synthesize a private property or explicitly declare an ivar then
explicitly override the getter instead.