Add missing cancellable to text input handler
Without this if a channel method call was in progress and then the handler
was destroyed the call could attempt to access the destroyed handler. This
is possible as the handler is recreated on engine restart.

However, in this case the result is not used so no invalid memory access
would occur. Adding the cancellable fixes any future change where this
would no longer be the case.