[Impeller] Run simulator tests with Impeller enabled.
The Info.plist override was missed earlier.

The main changes to the tests are due to the UI thread merge with the
platform thread. The amendment to the "spawn" API are due to the fact
that the assertions checked for the presence of GrDirectContext which
will never exist with Impeller.

Unblocks https://github.com/flutter/engine/pull/56706