Eliminate ShellTestPlatformView::BackendType::kDefaultBackendType
`kDefaultBackendType` is intended to make life easier for authors of
tests, but in any switch statement where it's used (currently just a
single location), we rely on ordering it first and `#ifdef`ing out all
backends that aren't available.

Instead, we define a static function that returns the default that
callers can invoke instead. This avoids relinace on case ordering and
fallthrough.

In https://github.com/flutter/engine/pull/56722 we split backends out
into separate translation units, and remove the `#ifdef`s which means we
can't rely on this trick anymore.