iOS: Migrate PlatformViewsController to Objective-C
This migrates PlatformViewController from C++ to Objective-C. Generally,
we try to keep the embedder interfaces and components written in
Objective-C except for the few places where C++ interfaces are requried
to interface with engine APIs such as Shell and PlatformView (e.g. the
PlatformViewIOS subclass). Now that the implementation is Objective-C,
the class and file are renamed to match Objective-C naming conventions.

This allows us to take advantage of ARC and weak references, which
eliminates the need for std::shared_ptr, fml::WeakPtr etc. Further, this
eliminates some particularly unintuitive behaviour wherein this class
was owned via a std::shared_ptr held by FlutterEngine, and injected into
many other classes (e.g. AccessibilityBridge) via a std::shared_ptr&
reference -- such that only one instance of the std::shared_ptr actually
ever existed, presumably to avoid std::shared_ptr refcounting overhead.
Given that this overhead was only incurred a single time at engine
initialisation, this seems like overkill. One might ask why it wasn't
therefore held in a `std::unique_ptr` and a `std::unique_ptr&` reference
passed around. Likely, this was because we wanted to take a
`fml::WeakPtr` reference on it.

Regardless, none of this is necessary any longer now that we can inject
`__weak FlutterPlatformViewsController*` instances to classes that use
it.

To be clear, this patch makes no attempt whatsoever to simplify or clean
up the interface or implementation of this class. This class ties
together far too many concepts and is injected into far too many places,
and we should break it up and simplify it. However, the goal of this
patch was simply to port to an Objective-C interface that plays nicely
with the rest of the iOS embedder. This does include a couple minor
cleanups in `#include`/`#import` order and usage to match our style
guide.