iOS: Eliminate non-handling of non-zero origin platformviews
In flutter/engine#35501, handling was added to log a debug message to
the console in the case where a platform view with a non-zero origin was
identified.

Unfortunately:
* In unopt builds, the first thing we do in that block is to call
  FML_DCHECK asserting that the origin is zero, so we never actually
  emit the log statement.
* In opt builds, FML_DCHECK is a no-op, so users are unlikely to
  actually ever notice the crash.

The proper fix is to eliminate this restriction, but in the meantime,
this eliminates this block entirely and leaves the TODO. We've had only
two comments on that bug in the 2.5 years since it was added.

Issue: https://github.com/flutter/flutter/issues/109700