iOS: Move global to static in platformviews controller
This is a minor refactoring that moves a global bool to a static bool in
PlatformViewsController. This doesn't improve the logic but does limit
the scope of the bool to a single method.

Also migrates from C++ style namespace { ... } to Obj-C style static
functions. These are entirely equivalent, both restrict these symbols to
the current translation unit.