iOS: Add null checks on shell dereference
`FlutterEngine` at the `_shell` unique_ptr ivar it owns have different
lifetimes. `_shell` is initialised transitively from
`runWithEntrypoint`, and reset in `[FlutterEngine destroyContext]`,
which is called transitively from `[FlutterviewController dealloc]` via
`[FlutterEngine notifyViewControllerDeallocated]`.

As such, all uses of `_shell` should be checked either via an assertion,
in cases we know the shell should be non-null, or via a runtime null
check in cases where it's expected that it may be null.

Specifically, this guards against a crash that can occur if we get a
CoreAnimation transaction commit callback for an inflight frame just as
we're shutting down the app (or removing the FlutterView in an
add-to-app scenario).

Example stack trace:
```
0  Flutter                        0x11b28 -[FlutterEngine platformView] + 53 (weak_ptr.h:53)
1  Flutter                        0x11994 -[FlutterEngine updateViewportMetrics:] + 186 (ref_ptr.h:186)
2  Flutter                        0x1f854 -[FlutterViewController updateViewportMetricsIfNeeded] + 427 (vector:427)
3  Flutter                        0x1f9b8 -[FlutterViewController viewDidLayoutSubviews] + 1411 (FlutterViewController.mm:1411)
4  UIKitCore                      0x8c864 -[UIView(CALayerDelegate) layoutSublayersOfLayer:] + 2376
5  QuartzCore                     0x1fa0c CA::Layer::layout_if_needed(CA::Transaction*) + 516
6  QuartzCore                     0x1ae84c CA::Context::commit_transaction(CA::Transaction*, double, double*) + 516
7  QuartzCore                     0x2888 CA::Transaction::commit() + 648
```

Issue: https://github.com/flutter/flutter/issues/159639