Make Felt unzip Chrome into a temporary directory next to the final location of the files
The Web Felt script downloads a Chrome zip archive, extracts it to a temporary directory, and then moves the files to a location expected by the script.

Previously Felt used a temporary directory based on io.Directory.systemTemp.  This may not work because the FileSystemEntity.rename operation used to move the files can fail if the system temp directory is located in a different filesystem from the destination directory.