Build SDKs for the DDC library bundle format
In order to support hot reload, we initially planned to use
the DDC module format. When that became infeasible, we introduced
the library bundle format.

This format can only be emitted when passed as the single module
arg to a DDC compile, so this change splits the generation of
SDKs into multiple steps.

We also do not support unsound null safety with the new format,
so it should never be generated.

When Flutter tools no longer tests for the DDC module format and
removes support for it, the generated for that format can be
deleted.